+++ tensorizer/stream_io.py
@@ -23,7 +23,7 @@ import redis
 import tensorizer._version as _version
 import tensorizer._wide_pipes as _wide_pipes
 
-__all__ = ["open_stream", "CURLStreamFile", "RedisStreamFile"]
+__all__ = ["open_stream", "CURLStreamFile", "RedisStreamFile", "CAInfo"]
 
 logger = logging.getLogger(__name__)
 
@@ -117,6 +117,75 @@ def _get_s3cfg_values(
     )
 
 
+class CAInfo:
+    """
+    Configuration class for cURL's SSL certificate verification behaviour.
+
+    See Also:
+        https://curl.se/docs/sslcerts.html
+    """
+
+    __slots__ = ("_curl_flags",)
+
+    def __init__(
+        self,
+        *,
+        cacert: Optional[str] = None,
+        capath: Optional[str] = None,
+        allow_untrusted: bool = False,
+    ):
+        """
+        Creates a configuration defining custom paths to use for SSL certificate
+        verification, or disables SSL certificate verification.
+
+        Exactly one of `cacert`, `capath`, or `allow_untrusted` must be
+        specified.
+        Use None instead of an instance of this class for cURL's default
+        certificate verification behaviour.
+
+        Args:
+            cacert: Path to a PEM file to use for certificates.
+                Corresponds to ``--cacert`` or ``CURLOPT_CAINFO`` in cURL.
+            capath: Path to a directory to scan for certificates.
+                Corresponds to ``--capath`` or ``CURLOPT_CAPATH`` in cURL.
+            allow_untrusted: Skip verification of SSL certificates entirely.
+                Corresponds to ``-k``/``--insecure``
+                or setting ``CURLOPT_SSL_VERIFYPEER`` to ``FALSE`` in cURL.
+        """
+        if (cacert, capath).count(None) + (not allow_untrusted) != 2:
+            raise ValueError(
+                "Exactly one of cacert, capath, or allow_untrusted=True"
+                " must be specified. For default certificate verification"
+                " behaviour, use certificate_handling=None instead of"
+                " an explicit CAInfo instance"
+            )
+        if cacert is not None:
+            if not os.path.isfile(cacert):
+                raise ValueError(
+                    "Invalid cacert file path, see"
+                    " https://curl.se/docs/manpage.html#--cacert"
+                )
+            self._curl_flags = ("--cacert", str(cacert))
+        elif capath is not None:
+            if not os.path.isdir(capath):
+                raise ValueError(
+                    "Invalid capath directory path, see"
+                    " https://curl.se/docs/manpage.html#--capath"
+                )
+            self._curl_flags = ("--capath", str(capath))
+        elif allow_untrusted:
+            self._curl_flags = ("--insecure",)
+        else:
+            raise RuntimeError()
+
+    @property
+    def curl_flags(self):
+        return self._curl_flags
+
+    def __hash__(self):
+        return hash(self._curl_flags)
+
+
 class CURLStreamFile:
     """
     CURLStreamFile implements a file-like object around an HTTP download, the
@@ -147,6 +216,7 @@ class CURLStreamFile:
         headers: Dict[str, Any] = None,
         *,
         buffer_size: Optional[int] = None,
+        certificate_handling: Optional[CAInfo] = None,
     ) -> None:
         if buffer_size is None:
             buffer_size = 16 << 20  # 16mb
@@ -159,8 +229,23 @@ class CURLStreamFile:
                 " and could not be found."
             )
 
+        self._optional_curl_flags = []
+        if certificate_handling is not None:
+            if isinstance(certificate_handling, CAInfo):
+                self._optional_curl_flags.extend(
+                    certificate_handling.curl_flags
+                )
+            else:
+                raise TypeError(
+                    "certificate_handling must either be None"
+                    " (for cURL's default SSL certificate verification),"
+                    " or a CAInfo object,"
+                    f" not {certificate_handling.__class__.__name__}"
+                )
+
         cmd = [
             curl_path,
+            *self._optional_curl_flags,
             "--header",
             "Accept-Encoding: identity",
             "--include",
@@ -202,8 +287,20 @@ class CURLStreamFile:
         self.http_response_latencies.append(resp_begin - popen_end)
 
         if not resp.startswith((b"HTTP/1.1 2", b"HTTP/2 2")):
+            # If the error was caused by an invalid response, display that
+            explanation = f": {resp.decode('utf-8')}" if resp.strip() else ""
+            if not explanation:
+                # Otherwise, check if cURL gave a meaningful exit code
+                return_code: Optional[int] = self._curl.poll()
+                if isinstance(return_code, int) and return_code != 0:
+                    explanation = (
+                        f": cURL exit code {return_code:d};"
+                        f" see https://curl.se/docs/manpage.html#{return_code}"
+                    )
+
             self.close()
-            raise IOError(f"Failed to open stream: {resp.decode('utf-8')}")
+            raise IOError("Failed to open stream" + explanation)
+
         # Read the rest of the header response and parse it.
         # noinspection PyTypeChecker
         self.response_headers = http.client.parse_headers(self._curl.stdout)
@@ -267,6 +364,7 @@ class CURLStreamFile:
         """
         args = [
             curl_path,
+            *self._optional_curl_flags,
             "-I",  # Only read headers
             "-XGET",  # Use a GET request (in case HEAD is not supported)
             "-A",  # Set a custom user agent
@@ -836,6 +934,7 @@ def s3_download(
     s3_endpoint: str = default_s3_read_endpoint,
     buffer_size: Optional[int] = None,
     force_http: bool = False,
+    certificate_handling: Optional[CAInfo] = None,
 ) -> CURLStreamFile:
     url = _s3_download_url(
         path_uri=path_uri,
@@ -845,7 +944,11 @@ def s3_download(
     )
     if force_http and url.lower().startswith("https://"):
         url = "http://" + url[8:]
-    return CURLStreamFile(url, buffer_size=buffer_size)
+    return CURLStreamFile(
+        url,
+        buffer_size=buffer_size,
+        certificate_handling=certificate_handling,
+    )
 
 
 def _infer_credentials(
@@ -996,6 +1099,8 @@ def open_stream(
     s3_config_path: Optional[Union[str, bytes, os.PathLike]] = None,
     buffer_size: Optional[int] = None,
     force_http: bool = False,
+    *,
+    certificate_handling: Optional[CAInfo] = None,
 ) -> Union[CURLStreamFile, RedisStreamFile, typing.BinaryIO]:
     """
     Open a file path, http(s):// URL, or s3:// URI.
@@ -1037,6 +1142,12 @@ def open_stream(
         force_http: If True, force the use of HTTP instead of HTTPS for
             S3 downloads. This will double the throughput, but at the cost
             of security.
+        certificate_handling: Customize handling of SSL CA certificates for
+            HTTPS and S3 downloads.
+            Pass None to use default certificate verification, or an instance of
+            `tensorizer.stream_io.CAInfo` to use a different CA bundle
+            or to disable certificate verification entirely.
+            This option is useful when working with self-signed certificates.
 
     Returns:
         An opened file-like object representing the target resource.
@@ -1076,6 +1187,15 @@ def open_stream(
                 "https://raw.githubusercontent.com/EleutherAI/gpt-neo/master/README.md"
             ) as stream:
                 print(stream.read(128))
+
+        Opening an https:// URI for reading,
+        while skipping SSL certificate verification::
+
+            with open_stream(
+                "https://127.0.0.1/my-model.tensors",
+                certificate_handling=CAInfo(allow_untrusted=True),
+            ) as stream:
+                print(stream.read(128))
     """
     if isinstance(path_uri, os.PathLike):
         path_uri = os.fspath(path_uri)
@@ -1090,7 +1210,11 @@ def open_stream(
             raise ValueError(
                 'Only the mode "rb" is valid when opening http(s):// streams.'
             )
-        return CURLStreamFile(path_uri, buffer_size=buffer_size)
+        return CURLStreamFile(
+            path_uri,
+            buffer_size=buffer_size,
+            certificate_handling=certificate_handling,
+        )
     elif scheme == "redis":
         if normalized_mode != "br":
             raise ValueError(
@@ -1170,6 +1294,7 @@ def open_stream(
                 s3_endpoint,
                 buffer_size=buffer_size,
                 force_http=force_http,
+                certificate_handling=certificate_handling,
             )
             if error_context:
                 curl_stream_file.register_error_context(error_context)